move comps to separate files
